Unit: Add skeleton support for the sprite-mode WAIT_FOR_BUILD.
The skeleton structure is added for a Unit to go into WAIT_FOR_BUILD mode and order any other units to vacate the construction site.
The function to order units out of the construction site does nothing yet, so it's just a placeholder.
Temporary support for savegames is included, by not modifying the savegame format and 'fixing' on load, but this needs to be improved later on by creating a new version of the savegame format.
There is no codepath yet that takes a unit into WAIT_FOR_BUILD mode.