Unit: Enter WAIT_FOR_BUILD mode when the construction site is obstructed.
When the building site is obstructed for a unit, it enters a the wait-for-build mode to wait for the construction site to clear.
While waiting it will check each day if it can start the construction.
The sprite's WAIT_FOR_BUILD mode exits after a certain time-out (e.g. after 7 days), after which the order is cancelled as normal.