Unit - Wait for Build: Prevent endless sprite Wait-for-Build <-> Idle looping by cancelling all actions if wait for build fails.
If action_para2 is not reset, then unit idle processing will call reactive_idle_action to reissue the build order, causing another wait for build.
This also fixes the quirk in the original where a failed build causes a unit to move a few steps;
this occurred because reactivate_idle_action issued another build_firm action on an occupied location, which becoming a move_to.