Wait for Build: Use one of the move sprite frames for the unit while waiting for build.
The unit no longer rotates (that was temporary, of course), but now uses the middle frame of its movement sprite.
It looks like it's stuck mid-walk, but it's unique enough and not too ridiculous to be used.
Maybe, someday, someone will feel like making a specific wait-for-build sprite (or, more generally, a wait sprite).