Fixed crash with weapons from a War Factory after nation surrender.
This commit fixes a crash in Sprite::cur_sprite_frame / SPRITE_ATTACK.
The crash occurs with weapons of war, which were in the build queue of a War Factory that changed nationality to a nation that doesn't have the technology for it.
The resulting weapon has its weapon technology level set erroneously to 0 upon completion, resulting in Unit::fix_attack_info to not fixup the weapon's attack info (and thus attack_count = 3, rather than 1).
This then allows cur_attack > 0, but because weapons such as a catapult have only one attack sprite, this causes cur_sprite_frame to crash.