Fryhtans: Changed attack curve, such that attacks happen more often when more Lairs are around.
The original code had one Fryhtan attack every 6 months, regardless of the number of lairs.
The new curve increases the number of attacks with the amount of Lairs, but is less than linear in the number of lairs.
At the default amount of 10 Lairs, there are 3 times as many attacks with this new curve than with the original, so once every 2 months.